Move manifesthandler to separate module
