+++ pkg/dockerregistry/server/manifesthandler/doc.go
@@ -1,0 +1,2 @@
+// Package manifesthandler defines a common set of operations on all versions of manifest schema.
+package manifesthandler
