+++ pkg/dockerregistry/server/manifesthandler/manifesthandler.go
@@ -1,4 +1,4 @@
-package server
+package manifesthandler
 
 import (
 	"fmt"
