+++ pkg/dockerregistry/server/manifesthandler/manifestschema1handler.go
@@ -1,4 +1,4 @@
-package server
+package manifesthandler
 
 import (
 	"fmt"
