+++ pkg/dockerregistry/server/manifesthandler/manifestschema2handler.go
@@ -1,4 +1,4 @@
-package server
+package manifesthandler
 
 import (
 	"errors"
